feat: validations in the uuids and email fields

+++ src/app/controllers/ContactController.js
@@ -1,4 +1,5 @@
 const ContactsRepository = require('../repositories/ContactsRepository');
+const isValidUUID = require('../utils/isValidUUID');
 
 class ContactController {
   async index(req, res) {
@@ -12,11 +13,16 @@ class ContactController {
   async show(req, res) {
     // Obter 1 registro
     const { id } = req.params;
+
+    if (!isValidUUID(id)) {
+      return res.status(400).json({ error: 'Invalid contact id' });
+    }
+
     const contact = await ContactsRepository.findById(id);
 
     if (!contact) {
       // 404: Not Found
-      return res.status(404).json({ error: 'User not found' });
+      return res.status(404).json({ error: 'Contact not found' });
     }
 
     res.json(contact);
@@ -32,17 +38,26 @@ class ContactController {
       return res.status(400).json({ error: 'Name is required' });
     }
 
-    const contactExists = await ContactsRepository.findByEmail(email);
+    if (category_id && !isValidUUID(category_id)) {
+      return res.status(400).json({ error: 'Invalid cateogry' });
+    }
 
-    if (contactExists) {
-      return res.status(400).json({ error: 'This e-mail is already in use' });
+    if (email) {
+      const contactExists = await ContactsRepository.findByEmail(email);
+
+      if (contactExists) {
+        return res.status(400).json({ error: 'This e-mail is already in use' });
+      }
     }
 
     const contact = await ContactsRepository.create({
-      name, email, phone, category_id,
+      name,
+      email: email || null,
+      phone,
+      category_id: category_id || null
     });
 
-    res.json(contact);
+    res.status(201).json(contact);
   }
 
   async update(req, res) {
@@ -53,24 +68,37 @@ class ContactController {
       name, email, phone, category_id,
     } = req.body;
 
-    const contactExists = await ContactsRepository.findById(id);
+    if (!isValidUUID(id)) {
+      return res.status(400).json({ error: 'Invalid contact id' });
+    }
 
-    if (!contactExists) {
-      return res.status(400).json({ error: 'User not found' });
+    if (category_id && !isValidUUID(category_id)) {
+      return res.status(400).json({ error: 'Invalid cateogry' });
     }
 
     if (!name) {
       return res.status(400).json({ error: 'Name is required' });
     }
 
-    const contactByEmail = await ContactsRepository.findByEmail(email);
+    const contactExists = await ContactsRepository.findById(id);
 
-    if (contactByEmail && contactByEmail.id !== id) {
-      return res.status(400).json({ error: 'This e-mail is already in use' });
+    if (!contactExists) {
+      return res.status(400).json({ error: 'Contact not found' });
+    }
+
+    if (email) {
+      const contactByEmail = await ContactsRepository.findByEmail(email);
+
+      if (contactByEmail && contactByEmail.id !== id) {
+        return res.status(400).json({ error: 'This e-mail is already in use' });
+      }
     }
 
     const contact = await ContactsRepository.update(id, {
-      name, email, phone, category_id,
+      name,
+      email: email || null,
+      phone,
+      category_id: category_id || null
     });
 
     res.json(contact);
@@ -79,6 +107,10 @@ class ContactController {
   async delete(req, res) {
     // Deletar 1 registro
     const { id } = req.params;
+
+    if (!isValidUUID(id)) {
+      return res.status(400).json({ error: 'Invalid contact id' });
+    }
 
     await ContactsRepository.delete(id);
